Minor polishing

+++ src/threadlist.cpp
@@ -598,8 +598,8 @@ void ThreadList::postRestart(void)
   /* If MTCP_RESTART_PAUSE set, sleep 15 seconds and allow gdb attach. */
   if (getenv("MTCP_RESTART_PAUSE")) {
     struct timespec delay = {15, 0}; /* 15 seconds */
-    printf("Pausing 15 seconds. Do:  gdb <PROGNAME> %d\n",
-    	   THREAD_REAL_TID());
+    printf("Pausing 15 seconds. Do:  gdb <PROGNAME> %ld\n",
+    	   (long)THREAD_REAL_TID());
     nanosleep(&delay, NULL);
   }
